Hotel Review Mining/Web App Deployment/api/requirements.txt
Flask==1.1.2
nltk==3.5
tensorflow-cpu
Flask_Cors==3.0.10
pandas==1.2.3
Flask_RESTful==0.3.8
numpy==1.19.2
gunicorn